requirements.txt
streamlit==1.39.0
google-generativeai
requests
beautifulsoup4
duckduckgo-search